Rewrite TimeController using a PriorityQueue
The aim of this rewrite is to avoid the linear search that was
previously used to find the next expiring timer. Instead, keep the
timers in a PriorityQueue. A Long keeps track of what time it is,
so the timers can be arranged in the queue according to the time at
which they are scheduled to run.